Создал несколько тестов для обработчика сервера с использованием библиотеки testify
